Applied DelegateInstallConvention to Terraform project
