+++ source/Calamari.Terraform/TerraformCommand.cs
@@ -1,9 +1,13 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.IO;
+using System.Threading.Tasks;
+using Calamari.Aws.Integration;
+using Calamari.Commands;
 using Calamari.Commands.Support;
 using Calamari.Deployment;
 using Calamari.Deployment.Conventions;
+using Calamari.Extensions;
 using Calamari.Integration.FileSystem;
 using Calamari.Integration.Packages;
 using Calamari.Integration.Processes;
@@ -15,15 +19,13 @@ namespace Calamari.Terraform
     {
         const string DefaultTerraformFileSubstitution = "**/*.tf\n**/*.tf.json\n**/*.tfvars\n**/*.tfvars.json";
 
-        private readonly IConvention step;
         readonly IVariables variables;
-        readonly ICalamariFileSystem fileSystem;
+        protected readonly ICalamariFileSystem fileSystem;
         private string packageFile;
 
 
-        protected TerraformCommand(IVariables variables, ICalamariFileSystem fileSystem, IConvention step)
+        protected TerraformCommand(IVariables variables, ICalamariFileSystem fileSystem)
         {
-            this.step = step;
             this.variables = variables;
             this.fileSystem = fileSystem;
             Options.Add("package=", "Path to the package to extract that contains the package.", v => packageFile = Path.GetFullPath(v));
@@ -56,7 +58,7 @@ namespace Calamari.Terraform
                 new SubstituteInFilesConvention(fileSystem, substituter,
                     _ => true,
                     _ => FileTargetFactory(runAutomaticFileSubstitution ? DefaultTerraformFileSubstitution : string.Empty, additionalFileSubstitution)),
-                step
+                new DelegateInstallConvention(d => InstallAsync(d).GetAwaiter().GetResult())
             };
 
             var deployment = new RunningDeployment(packageFile, variables);
@@ -74,6 +76,68 @@ namespace Calamari.Terraform
                                             ? string.Empty
                                             : "\n" + additionalFileSubstitution))
                 .Split(new[] {"\r", "\n"}, StringSplitOptions.RemoveEmptyEntries);
+        }
+        
+        
+        async Task InstallAsync(RunningDeployment deployment)
+        {
+            var variables = deployment.Variables;
+            var environmentVariables = new Dictionary<string, string>();
+            var useAWSAccount = variables.Get(TerraformSpecialVariables.Action.Terraform.AWSManagedAccount, "None") == "AWS";
+            var useAzureAccount = variables.GetFlag(TerraformSpecialVariables.Action.Terraform.AzureManagedAccount);
+
+            if (useAWSAccount)
+            {
+                var awsEnvironmentGeneration = await AwsEnvironmentGeneration.Create(variables).ConfigureAwait(false);
+                environmentVariables.MergeDictionaries(awsEnvironmentGeneration.EnvironmentVars);
+            }
+
+            if (useAzureAccount)
+            {
+                environmentVariables.MergeDictionaries(AzureEnvironmentVariables(variables));
+            }
+
+            Execute(deployment, environmentVariables);
+        }
+        
+        protected abstract void Execute(RunningDeployment deployment, Dictionary<string, string> environmentVariables);
+
+        // We not referencing the Azure project because we want to be multi-platform 
+        static Dictionary<string, string> AzureEnvironmentVariables(IVariables variables)
+        {
+            string AzureEnvironment(string s)
+            {
+                switch (s)
+                {
+                    case "AzureChinaCloud":
+                        return "china";
+                    case "AzureGermanCloud":
+                        return "german";
+                    case "AzureUSGovernment":
+                        return "usgovernment";
+                    default:
+                        return "public";
+                }
+            }
+
+            var environmentName = AzureEnvironment(variables.Get(SpecialVariables.Action.Azure.Environment));
+
+            var account = variables.Get(SpecialVariables.Action.Azure.AccountVariable)?.Trim();
+            var subscriptionId = variables.Get($"{account}.SubscriptionNumber")?.Trim() ?? variables.Get(SpecialVariables.Action.Azure.SubscriptionId)?.Trim();
+            var clientId = variables.Get($"{account}.Client")?.Trim() ?? variables.Get(SpecialVariables.Action.Azure.ClientId)?.Trim();
+            var clientSecret = variables.Get($"{account}.Password")?.Trim() ?? variables.Get(SpecialVariables.Action.Azure.Password)?.Trim();
+            var tenantId = variables.Get($"{account}.TenantId")?.Trim() ?? variables.Get(SpecialVariables.Action.Azure.TenantId)?.Trim();
+            
+            var env = new Dictionary<string, string>
+            {
+                {"ARM_SUBSCRIPTION_ID", subscriptionId},
+                {"ARM_CLIENT_ID", clientId},
+                {"ARM_CLIENT_SECRET", clientSecret},
+                {"ARM_TENANT_ID", tenantId},
+                {"ARM_ENVIRONMENT", environmentName}
+            };
+
+            return env;
         }
     }
 }
